// src/services/api/gen/types.gen.ts
export type ApprovalStatusFe1Enum = 'approved' | 'rejected' | 'pending' | 'need_information' | 'kyc_pending';
// ...
export type BusinessAdminList = {
    readonly id: string;
    name: string;
    approval_status: ApprovalStatusFe1Enum;
    readonly first_name: string;
    readonly last_name: string;
    readonly remarks: Array<BusinessAdminListAction>;
};

export type BusinessAdminListAction = {
    action: string;
    remark?: string | null;
    readonly added_by_first_name: string;
    readonly added_by_last_name: string;
    readonly created_at: string;
};